[PL] moved parameter related files to subdirectory
